+++ examples/unpaced_mode.js
@@ -1,8 +1,8 @@
 var hypertimer = require('../index');
 
-// create a hypertimer running discrete events,
-// jumping from event to the next event.
-var timer = hypertimer({rate: 'discrete-event'});
+// create a hypertimer running discrete events as fast as possible
+// in unpaced mode, jumping from event to the next event.
+var timer = hypertimer({paced: false});
 
 timer.setTimeout(function () {
   console.log('Timeout A');
@@ -21,7 +21,7 @@ timer.setTimeout(function () {
 
 }, 10000);
 
-// Because of running discrete events, the application will finish immediately
+// Because of running in unpaced mode, the application will finish immediately
 // and all events are executed in deterministic order.
 //
 // Output will be: